+++ perlin_noise.py
@@ -17,6 +17,7 @@ def maxItemLength(a):
             maxLen = max(maxLen, len(repr(a[row][col])))
     return maxLen
 
+# https://www.cs.cmu.edu/~112/notes/notes-2d-lists.html#printing
 def print2dList(a):
     if (a == []):
         # So we don't crash accessing a[0]
@@ -33,23 +34,33 @@ def print2dList(a):
         print(' ]')
     print(']')
 
+# returns unit vector between two points
 def unitVector(x, y):
     k = (x ** 2 + y ** 2) ** -0.5
     return (k * x, k * y)
 
+# returns random gradient vector
 def gradient():
     x, y = random.uniform(-1, 1), random.uniform(-1, 1)
     return unitVector(x, y)
 
+# smoothening function
 def fade(x):
     return 6 * x ** 5 - 15 * x ** 4 + 10 * x ** 3
 
+# returns dot product of two vectors
 def dot(v1, v2):
     return v1[0] * v2[0] + v1[1] * v2[1]
 
+# linear interpolation formula
 def interpolate(d, a, b):
     return (1 - d) * a + d * b
 
+# elliptic parabaloid function to create cloud shape
+def f(x, y, width, height, dx, dy, a, level):
+    return ((a * (x - dx * width) / width) ** 2 + ((y - dy * height) / height) ** 2) * level
+
+# creates noise for point x,y
 def noise(x, y, scale):
     # corners
     p1 = int(x), int(y)
@@ -68,13 +79,11 @@ def noise(x, y, scale):
     bottom = interpolate(dx, dot(g3, d3), dot(g4, d4))
     return interpolate(dy, top, bottom) * scale
 
-width, height = 166, 102
-result = [[0] * (width) for y in range(height)]
-
-octaves = 2
-
-def octave(result, per, lac):
+# returns 2D list of noise with octave, persistance, lacunarity parameters
+def octave(result, per, lac, octaves, level):
+    #! random eccentricity for elliptic base
     a = random.uniform(0.5, 2)
+    #! position of cloud center
     dx, dy = random.random(), random.random()
     for y in range(len(result)):
         for x in range(len(result[0])):
@@ -83,21 +92,11 @@ def octave(result, per, lac):
             noiseH = 0
             for i in range(octaves):
                 noiseH += (2 * noise(x * freq, y * freq, 1) + 1) * f(x, y, len(result[0]), 
-                            len(result), dx, dy, a)
+                            len(result), dx, dy, a, level)
                 amp *= per
                 freq *= lac
             result[y][x] = float("{:.2f}".format(noiseH))
     return result
 
-def f(x, y, width, height, dx, dy, a):
-    return ((a * (x - dx * width) / width) ** 2 + ((y - dy * height) / height) ** 2)
-
-def Noise2D(L):
-    for y in range(len(L)):
-        for x in range(len(L[0])):
-            L[y][x] = (noise(x / 100, y / 100, 1) + 1) / 2
-    return L 
-
-result = octave(result, 0.4, 1)
-#print2dList(result)
-#result = octave(result, 10, 4000)
+width, height = 166, 102
+result = octave([[0] * (width) for y in range(height)], 0.4, 1, 2, random.randint(1,20))